Create Header folder and moved relevant files to this folder
