+++ sudoku-solver/src/Components/App.js
@@ -1,7 +1,7 @@
 import "../css/App.css";
 import { SudokuBoard } from "./SudokuBoard";
 import { BoardUse } from "./BoardUse";
-import { Header } from "./Header";
+import { Header } from "./Header/Header";
 
 function App() {
   return (
